Fix purging problem when providing sitemap url.
Purging is based on the site base url (the first argument for Site):
the documents in Solr are identified by looking if they start with
this site base url.

By making it possible to provide a sitemap url as first argument to
Site, the purging broke, because the indexed document urls do not
start with the sitemap (!) url.

This change removes the functionality to provide a sitemap url
directly as first argument.
Instead, the sitemap_urls keyword argument must be used, where a list
of sitemap urls is provided.

All existing configurations with a sitemap url configured explicitly
must be updated when updating ftw.crawler.
For the affected sites the administrator should purge the solr index
manually and reindex.